requirements.txt
Flask==3.0.3
Flask_Bootstrap==3.3.7.1
flask_wtf==1.2.1
pymongo==4.7.3
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
